Add files via upload
This is the first attempt to make a worm that goes through the 192.168.0.0 subnet, 10.0.0.0 subnet, and the 172.16.0.0 subnet.
The worm finds vulnerable targets with the Nmap module. The user variable is set to 'test' which later will be set to 'admin' after I am finished completing the worm. It then runs through the pass.txt to use a dictionary attack. Then it spreads through the ssh connection.
I am open for any help in completing my worm, or any advice in simplifying my coding. I will later add a way to make a back door with the worm that will later be controlled by a botnet. Thank you for your advice and support!
Side note: I am going to add the botnet later when I figure out the worm and any other programs i have to make in order for the worm to work.